Changed indentity scope return to long/bigint.

+++ PPOk Notification System/PPOk Notifications/Service/SQL/User_Insert.sql
@@ -16,4 +16,4 @@ VALUES
 	@Phone,
 	1
 )
-SELECT CAST(SCOPE_IDENTITY() as int)
+SELECT CAST(SCOPE_IDENTITY() as bigint)
